+++ frontend/src/business/components/track/issue/TestCaseIssueList.vue
@@ -60,7 +60,7 @@
 
 <script>
 import MsTable from "@/business/components/common/components/table/MsTable";
-import MsTableColumn from "@/business/components/common/components/table/Ms-table-column";
+import MsTableColumn from "@/business/components/common/components/table/MsTableColumn";
 import PriorityTableItem from "@/business/components/track/common/tableItems/planview/PriorityTableItem";
 import TypeTableItem from "@/business/components/track/common/tableItems/planview/TypeTableItem";
 import TestCaseRelateList from "@/business/components/track/issue/TestCaseRelateList";
